* react-messaging: (27 commits)
  complete single chat
  complete real-time credits
  complete init chat
  ulti init and complete payment generator
  change formula of charge money
  parseFloat for all product price
  add link into order_status
  fixed minor bug when payment by credits cards
  add console.log for step for chargeCard function. ParseFloat for subtotal and total
  install twilio and start call support button
  change server key from public to secret
  signal when new message coming
  completed basic chat features
  fixed position of chat panel
  fixed number and real number in shopping cart
  fixed float with 2 decimal number
  add kitchenlist to search result
  replace publishable live Stripe API key
  style for message panel
  handle keyword
  ...

+++ client/Components/shopping_cart/shopping_cart.js
@@ -40,7 +40,7 @@ Template.sc_cost_summary.helpers({
   'total_food_price': function () {
     var total_food_price = 0;
     Shopping_cart.find({ "buyer_id": Meteor.userId() }).map(function (doc) {
-      total_food_price += parseInt(doc.total_price_per_dish);
+      total_food_price += parseFloat(doc.total_price_per_dish);
     });
     return total_food_price;
   },
@@ -61,7 +61,7 @@ Template.sc_cost_summary.helpers({
     var total_price = 0
 
     Shopping_cart.find({ "buyer_id": Meteor.userId() }).map(function (doc) {
-      total_food_price += parseInt(doc.total_price_per_dish);
+      total_food_price += parseFloat(doc.total_price_per_dish);
     });
 
     total_price = no_destination * delivery_cost_per_place + total_food_price
